requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiohttp==3.10.5
tenacity==8.2.2
requests==2.32.3
python-dotenv==1.0.1
pytest-asyncio==0.22.0
aioresponses==0.7.8
python-json-logger==2.0.7
